requirements.txt
git+https://github.com/k17-vp-1752001/EmoteCounterBot.git
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
discord.py